pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel", "pip-tools>=7.4"]
build-backend = "setuptools.build_meta"

[project]
name = "nautilus-hmm-binance"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "nautilus_trader[binance]>=1.210.0",
  "fastapi==0.120.4",
  "uvicorn[standard]==0.30.6",
  "httpx==0.27.2",
  "requests==2.32.4",
  "jinja2==3.1.6",
  "prometheus-client==0.20.0",
  "prometheus-fastapi-instrumentator==6.1.0",
  "pandas==2.2.3",
  "numpy==1.26.4",
  "pyarrow>=20.0.0,<22.0.0",
  "scikit-learn==1.5.2",
  "xgboost==2.1.1",
  "hmmlearn==0.3.2",
  "duckdb==1.1.0",
  "python-dotenv==1.0.1",
  "loguru==0.7.2",
  "pydantic==2.8.2",
  "pydantic-settings==2.11.0",
  "pyjwt[crypto]==2.9.0",
  "PyYAML==6.0.1",
  "tabulate==0.9.0",
  "matplotlib==3.8.4",
  "watchfiles==0.21.0",
  "watchdog==4.0.1"
]

[project.optional-dependencies]
dev = [
  "bandit==1.7.9",
  "black==24.10.0",
  "hypothesis==6.115.5",
  "mypy==1.11.2",
  "pip-audit==2.7.3",
  "pip-tools==7.4.1",
  "pre-commit==3.8.0",
  "pytest==8.3.3",
  "pytest-asyncio==0.23.8",
  "respx[httpx]==0.21.1",
  "trio==0.26.2",
  "ruff==0.6.9",
  "types-python-dateutil==2.9.0.20240906",
  "types-requests==2.32.0.20241016",
  "web3==6.20.1",
  "pandas-stubs==2.2.2.240807",
  "eth-typing>=4.2.0",
  "setuptools==80.9.0"
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"

[tool.black]
line-length = 100
target-version = ["py311"]
exclude = '''
/(
  \.git
| \.mypy_cache
| \.pytest_cache
| build
| dist
| frontend/node_modules
| frontend/build
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = [
  "scripts/**",
  "build/",
  "dist/",
  ".venv/",
  "data/",
  "docs/**/generated_*",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "TRY", "BLE", "UP", "PLR", "N"]
ignore = ["E203", "E501"]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "TRY", "BLE", "B", "PLR", "N", "UP"]
"frontend/**" = ["S101", "E501"]
"engine/legacy_*.py" = ["TRY", "BLE", "B", "PLR", "N", "UP"]  # TODO: remove by 2025-12-31
"tools/burst_signals.py" = ["TRY"]  # TODO: remove by 2025-12-31 legacy CLI

[tool.ruff.lint.isort]
known-first-party = ["engine", "ops", "services"]
force-single-line = false

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false
namespace_packages = true
explicit_package_bases = true
exclude = [
  "frontend/",
  "docs/",
  "scripts/",
  "tests/e2e/",
  "services/backtest_suite/",
  "services/backtester/",
]

[[tool.mypy.overrides]]
module = ["ccxt"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["engine.app"]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["services.*.app.main"]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["ops.ui_api"]
disallow_untyped_defs = true

[tool.bandit]
skips = ["B101"]
exclude_dirs = ["tests", "frontend", "docs"]
targets = ["engine", "ops", "services", "shared", "ml_service"]

[tool.pytest.ini_options]
minversion = "8.0"
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-q --maxfail=1 --disable-warnings -p no:web3.tools.pytest_ethereum"
